split layouts; move layout for authenticated pages to (app) group
