+++ librefit-web/src/routes/(app)/tracker/weight/+page.svelte
