+++ librefit-web/src/routes/(app)/tracker/weight/+server.js
