+++ librefit-web/src/routes/+error.svelte
@@ -8,7 +8,7 @@
     <div class="container mx-auto p-8 space-y-8">
         <h1>{$page.status}: {$page.error.message}</h1>
         <p>Whoops. Seems like that took a wrong turn.</p>
-        <button class="btn variant-filled-primary" on:click|preventDefault={goto('/')}></button>
+        <button class="btn variant-filled-primary" on:click|preventDefault={goto('/')}>Main Page</button>
         <div class="flex justify-center">
             <ErrorBot height="60vh"/>
         </div>
